6.1.0版本

+++ app/src/main/java/com/hr/ui/ui/resume/activity/ResumeProfessionSkillActivity.java
@@ -332,7 +332,7 @@ public class ResumeProfessionSkillActivity extends BaseActivity<ResumeProfession
             ToastUitl.showShort("请填写使用时间");
             return;
         }
-        if(Integer.parseInt(etResumeProjectExpUseTime.getText().toString())==0){
+        if(Double.parseDouble(etResumeProjectExpUseTime.getText().toString())==0){
             ToastUitl.showShort("使用时间必须大于0");
             return;
         }
